Further reduce the occurrences of SWIGTYPE_p unresolved proxy types. This submission handles class methods of the following signature:
VALUE_TYPE* MethodName(FdoInt32& count);

For such methods we define a specialized collection class of VALUE_TYPE, we then %extend the class with an alternative method that wraps the original method, to return a FdoBasicValueCollection<VALUE_TYPE> and the %ignore the original method so the SWIGTYPE_p unresolved type does not leak out. This results in a more friendly API in C# as they now deal with a collection instead of an opaque pointer to a C-array.